monitoring_agent/requirements.txt
psycopg2-binary
dotenv
python-dotenv
flask
